reworked Trace structure
this is a test implementation to make trace reusable. Trace now is a
base class, two classes are derived from it: TraceSerial, TraceErr.
This shows how the interface stays the same, but the trace object
changes. In the real Esp project this will either point to a serial
implementation or an mqtt  implementation.