TC036: DHT refactoring and Alarm mode in PIR
This version introduces the alram handling in the neo pixels which are controlled by mqtt with a broadcast command. This generates an police light. The light is triggered every 500msec and showed some delays. The delays are generated by delay function calls in the main loop. In the DHT this comes from the measurement state handling and this was removed now.